Replace redux-promises with redux-thunk and make appropriate changes
Affects all actions and all reducers. Adds a bit more length to those
source files but also provides an enormous amount of flexibility in the
future for more complex actions. Significantly reduces the complexity
of the reducers and allows us to compartmentalize certain reducers
since we can call them in chain.

Does help with some complexity in connected components.