Running unit tests through npm as pre-commit hook is antipattern
In particular, it isn't os-independent (it fails on windows and normal
work arounds fail because of the complex nature of phpunit's cli usage).
Pre-commit hooks are the more significant issue in general. They
discourage frequent commits and encourage undesirable developer
behaviors. Developers should be comfortable and familiar with running
unit tests and running them often.